backend/requirements.txt
Django>=5.2.5
Pillow>=10.0.0
django-cors-headers>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.56.0